feat(#3): 어드민 검증 로직 추가 - 해당 채팅방의 어드민인지 검증하도록 로직을 추가했어요.

+++ src/main/java/com/teka/application/chatroom/service/DeleteUserService.java
@@ -1,7 +1,11 @@
 package com.teka.application.chatroom.service;
 
+import com.teka.application.chatroom.exception.ChatRoomAdminMismatchException;
 import com.teka.application.chatroom.port.in.DeleteUserUseCase;
+import com.teka.application.chatroom.port.out.CheckAdminPort;
 import com.teka.application.user.port.out.DeleteUserPort;
+import com.teka.domain.admin.AdminId;
+import com.teka.domain.chatroom.ChatRoomId;
 import com.teka.domain.user.UserId;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -13,9 +17,13 @@ import java.util.List;
 public class DeleteUserService implements DeleteUserUseCase {
 
     private final DeleteUserPort deleteUserPort;
+    private final CheckAdminPort checkAdminPort;
 
     @Override
-    public void execute(List<UserId> userIdList) {
+    public void execute(List<UserId> userIdList, ChatRoomId chatRoomId, AdminId adminId) {
+        if (!checkAdminPort.checkChatRoomByAdminId(chatRoomId, adminId)) {
+            throw new ChatRoomAdminMismatchException();
+        }
         userIdList.forEach(deleteUserPort::deleteByUserId);
     }
 }
